% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtime_models.R
\name{benchmark_odin}
\alias{benchmark_odin}
\title{Benchmark Odin metatpopulation model}
\usage{
benchmark_odin(model, times, nRun)
}
\arguments{
\item{model}{Odin model}

\item{times}{the time vector}

\item{nRun}{the number of runs}
}
\value{
the result of the benchmark
}
\description{
Benchmark Odin metatpopulation model
}
