% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSimInfModel.R, R/runtime_models.R
\name{benchmark_siminf}
\alias{benchmark_siminf}
\title{Benchmark a siminf model}
\usage{
benchmark_siminf(model, nRun)

benchmark_siminf(model, nRun)
}
\arguments{
\item{model}{a SimInf model}

\item{nRun}{the number of runs}
}
\value{
the results of the benchmarck

results of the benchmark
}
\description{
Benchmark a siminf model

Benchmark SimInf metapopulation model
}
