% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siminf_supportFunctions.R
\name{convert_siminfU}
\alias{convert_siminfU}
\title{Converts the output of siminf to a long data frame}
\usage{
convert_siminfU(U, times, comparts, nmetapop)
}
\arguments{
\item{U}{the U object from the model output}

\item{times}{vector of time steps}

\item{comparts}{vector of compartment names e.g. S and I}

\item{nmetapop}{number of metapopulations}
}
\value{
data frame with columns for time, metapopulation, state compartment and the value of the variable
}
\description{
Converts the output of siminf to a long data frame
}
