% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSimInfModel.R
\name{getSimInfModel}
\alias{getSimInfModel}
\title{Create the SimInf model}
\usage{
getSimInfModel(tspan, npop, n_initial_infected, starting_pop, trans_mat, beta)
}
\arguments{
\item{tspan}{the time vector}

\item{npop}{the number of subpopulation}

\item{n_initial_infected}{the number of initially infected individuals}

\item{starting_pop}{the initial state}

\item{trans_mat}{the transfer matrix}

\item{beta}{the probability of infection}
}
\value{
a SimInf model
}
\description{
Create the SimInf model
}
