% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odin_ouput_functions.R
\name{getTransitionMatrix}
\alias{getTransitionMatrix}
\title{Rearrange odin output matrix into a matrix of transfers of infected individuals}
\usage{
getTransitionMatrix(sim, n_subpop)
}
\arguments{
\item{sim}{Raw odin output matrix}
}
\value{
Square matrix of transfers of infected individuals
}
\description{
Rearrange odin output matrix into a matrix of transfers of infected individuals
}
