% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_state_randomisation.R
\name{initial_state_randomisation}
\alias{initial_state_randomisation}
\title{Randomisation of the initial number of infected individuals}
\usage{
initial_state_randomisation(
  size_subpop,
  I_initial_time,
  initial_state_rando_proc,
  transfer_matrix
)
}
\arguments{
\item{size_subpop}{the size of the subpopulations}

\item{I_initial_time}{the initially infected in each subpopulation or the total number}

\item{initial_state_rando_proc}{the procedure of initial state randomisation if I_initial_time has length 1}

\item{transfer_matrix}{the transfer matrix}
}
\value{
List with the vector of the subpopulation sizes and the vector of the initial
number of infected individuals by subpopulation
}
\description{
Randomisation of the initial number of infected individuals
}
