% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_odin.R
\name{initialize_odin_binomial}
\alias{initialize_odin_binomial}
\title{Initialize the Odin model}
\usage{
initialize_odin_binomial(
  beta,
  n_subpop,
  size_subpop,
  transfer_matrix,
  I_per_subpop,
  time_step
)
}
\arguments{
\item{beta}{the probability of infection}

\item{n_subpop}{number of subpopulation}

\item{size_subpop}{the size of the subpopulations}

\item{transfer_matrix}{the transfer matrix}

\item{I_per_subpop}{the initially infected in each subpopulation}
}
\value{
an Odin model
}
\description{
Initialize the Odin model
}
