% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_odin.R
\name{initialize_sis}
\alias{initialize_sis}
\title{Initialize the Odin model}
\usage{
initialize_sis(
  beta,
  alpha,
  size_subpop,
  transfer_matrix,
  I_initial_time,
  community_prev,
  initial_state_rando_proc = "none"
)
}
\arguments{
\item{beta}{the probability of infection}

\item{alpha}{the recovery rate}

\item{size_subpop}{the size of the subpopulations}

\item{transfer_matrix}{the transfer matrix}

\item{I_initial_time}{the initially infected in each subpopulation or the total number}

\item{community_prev}{the community prevalence}

\item{initial_state_rando_proc}{the procedure of initial state randomisation if
I_initial_time has length 1: "none" is the default, "uniform", "largest_pop", "most_connected_pop"}
}
\value{
an Odin model
}
\description{
Initialize the Odin model
}
