% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siminf_supportFunctions.R
\name{make_fake_matrix}
\alias{make_fake_matrix}
\title{Creates a random matrix of transfers between metapopulations}
\usage{
make_fake_matrix(nmetapop, scale = 5)
}
\arguments{
\item{nmetapop}{number of metapopulations}

\item{scale}{average number of transfers for each pair of metapopulations}
}
\value{
transfer matrix
}
\description{
Creates a random matrix of transfers between metapopulations
}
