% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siminf_supportFunctions.R
\name{make_siminf_events}
\alias{make_siminf_events}
\title{Converts a daily transfer matrix into a data frame of daily transfer events}
\usage{
make_siminf_events(times, nmetapop, transfer_matrix, select = 1, shift = 0)
}
\arguments{
\item{times}{vector of time steps}

\item{nmetapop}{number of metapopulations}

\item{transfer_matrix}{matrix of daily transfers, from the row to the column}

\item{select}{the state compartments in the source metapopulation which can be sampled: by default any state can be chosen at random}

\item{shift}{ignore for now}
}
\value{
data frame of events
}
\description{
Converts a daily transfer matrix into a data frame of daily transfer events
}
