% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_state_randomisation.R
\name{rando_largest_pop}
\alias{rando_largest_pop}
\title{Randomisation among the largest subpopulations}
\usage{
rando_largest_pop(size_subpop, I_initial_time)
}
\arguments{
\item{size_subpop}{the size of the subpopulations}

\item{I_initial_time}{the initially infected in each subpopulation or the total number}
}
\value{
List with the vector of the subpopulation sizes and the vector of the initial
number of infected individuals by subpopulation
}
\description{
Randomisation among the largest subpopulations
}
