% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_state_randomisation.R
\name{rando_most_connected_pop}
\alias{rando_most_connected_pop}
\title{Randomisation among the most connected subpopulations}
\usage{
rando_most_connected_pop(size_subpop, I_initial_time, transfer_matrix)
}
\arguments{
\item{size_subpop}{the size of the subpopulations}

\item{I_initial_time}{the initially infected in each subpopulation or the total number}

\item{transfer_matrix}{the transfer matrix}
}
\value{
List with the vector of the subpopulation sizes and the vector of the initial
number of infected individuals by subpopulation
}
\description{
Randomisation among the most connected subpopulations
}
