% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_state_randomisation.R
\name{rando_uniform}
\alias{rando_uniform}
\title{Uniform randomisation procedure}
\usage{
rando_uniform(size_subpop, I_initial_time)
}
\arguments{
\item{size_subpop}{the size of the subpopulations}

\item{I_initial_time}{the initially infected in each subpopulation or the total number}
}
\value{
List with the vector of the subpopulation sizes and the vector of the initial
number of infected individuals by subpopulation
}
\description{
Uniform randomisation procedure
}
