% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_odin.R
\name{run_simulation}
\alias{run_simulation}
\title{Run initialized odin model and return list object}
\usage{
run_simulation(initialized_model, t_max, time_step, replicate)
}
\arguments{
\item{initialized_model}{Initialized odin model}

\item{t_max}{Last time step}

\item{time_step}{Duration of the time step}

\item{replicate}{Number of simulations}
}
\value{
List containing
\enumerate{
\item the prevalence and incidence per timestep, subpopulation, and simulation
\item the incidence per timestep, subpopulation, and simulation
\item the number of transferred infected individuals between every pair of subpopulation per timestep and simulation
\item the transfer matrix
\item the vector of subpopulation sizes
}
}
\description{
Run initialized odin model and return list object
}
